Assignments/session-1/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
idna==3.4
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
